# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "governed-bi API",
    "summary": "Governed NL2SQL serve flow + corpus/schema/audit, as JSON.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/schema": {
      "get": {
        "tags": [
          "schema"
        ],
        "summary": "Schema",
        "description": "Every table with its columns (types, roles, governance flags).\n\nParam-less this is the full dump (backward-compatible). ``schema`` filters\nto one namespace; ``limit``/``offset`` paginate (default: all rows, offset 0).",
        "operationId": "schema_schema_get",
        "parameters": [
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one schema namespace",
              "title": "Schema"
            },
            "description": "Filter to one schema namespace"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableResponse"
                  },
                  "title": "Response Schema Schema Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schema/summary": {
      "get": {
        "tags": [
          "schema"
        ],
        "summary": "Schema Summary",
        "description": "Lean catalog for the virtualized table list + the client search index.\n\nHeavy fields (sample_values, evidence, description) are dropped; fetch full\ndetail lazily via ``/schema/{table_id}``. ``schema`` filters to one\nnamespace; ``limit``/``offset`` paginate (default: all rows, offset 0);\n``total`` is the count BEFORE pagination.",
        "operationId": "schema_summary_schema_summary_get",
        "parameters": [
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one schema namespace",
              "title": "Schema"
            },
            "description": "Filter to one schema namespace"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schema/{table_id}": {
      "get": {
        "tags": [
          "schema"
        ],
        "summary": "Schema Table",
        "description": "Full detail for one table by asset id (404 when the id is unknown).",
        "operationId": "schema_table_schema__table_id__get",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/columns/{column_id}/related": {
      "get": {
        "tags": [
          "schema"
        ],
        "summary": "Column Related",
        "description": "Every semantic-layer item that touches one physical column (handoff \u00a714).\n\n``column_id`` is the derived id ``col_<table>_<physical_name>``. Returns\nterms binding it, rules scoping it, FK in/out, joins whose predicate touches\nit (resolved server-side), and metrics on its table (table-grain only).\n``404`` when the id does not resolve to a known column.",
        "operationId": "column_related_columns__column_id__related_get",
        "parameters": [
          {
            "name": "column_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Column Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnRelatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph": {
      "get": {
        "tags": [
          "schema"
        ],
        "summary": "Graph",
        "description": "Table-relationship graph for the ER view (nodes + join edges).\n\nOptional D15 scope: ``schema`` / ``focus`` / ``radius`` / ``node_budget``.\nWhen scoped, the response includes ``boundary`` (cross-schema stubs) and\n``meta`` (truncation + echoed scope). Param-less = full graph.",
        "operationId": "graph_graph_get",
        "parameters": [
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one schema namespace",
              "title": "Schema"
            },
            "description": "Filter to one schema namespace"
          },
          {
            "name": "focus",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Focus table asset id for a neighborhood",
              "title": "Focus"
            },
            "description": "Focus table asset id for a neighborhood"
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "BFS hops from focus (default 1)",
              "title": "Radius"
            },
            "description": "BFS hops from focus (default 1)"
          },
          {
            "name": "node_budget",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Max nodes to return (capped)",
              "title": "Node Budget"
            },
            "description": "Max nodes to return (capped)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/corpus/assets": {
      "get": {
        "tags": [
          "corpus"
        ],
        "summary": "Corpus Assets",
        "description": "Non-table assets (metrics/terms/joins/rules/few-shots/negatives).",
        "operationId": "corpus_assets_corpus_assets_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "join",
                    "metric",
                    "term",
                    "rule",
                    "few_shot",
                    "negative_example"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetRowResponse"
                  },
                  "title": "Response Corpus Assets Corpus Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/corpus/edit": {
      "post": {
        "tags": [
          "corpus"
        ],
        "summary": "Corpus Edit",
        "description": "Validate a corpus asset and, in dev, write it to the YAML tree.\n\nGated on ``capabilities.can_edit`` (403 otherwise). The asset is schema-\nvalidated (422 on a bad shape) then reference-checked against the rest of\nthe corpus; findings block the write and are returned with the diff so the\neditor can fix them. Prod PR mode is deferred; the request shape is stable.",
        "operationId": "corpus_edit_corpus_edit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchemaGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SchemaGraphNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SchemaGraphEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          },
          "boundary": {
            "items": {
              "$ref": "#/components/schemas/BoundaryEdgeResponse"
            },
            "type": "array",
            "title": "Boundary"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphMetaResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "SchemaGraphResponse"
      },
      "SchemaSummaryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TableSummaryResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "SchemaSummaryResponse"
      },
...
    }
  }
}